isGB2312 {tmcn}	R Documentation

Indicate whether the encoding of input string is GB2312.

Description

Indicate whether the encoding of input string is GB2312.

Usage

isGB2312(string, combine = FALSE)

Arguments

string
A character vector.

combine
Whether to combine all the strings.

Value

Logical value.

Author(s)

Jian Li <rweibo@sina.com>

Examples

isGB2312("hello")

[Package tmcn version 0.2-12 Index]