left {tmcn}	R Documentation

Extract the left or right substrings in a character vector.

Description

Extract the left or right substrings in a character vector.

Usage

left(string, n)
right(string, n)

Arguments

string
A character vector.

n
How many characters.

Value

A character vector.

Author(s)

Jian Li <rweibo@sina.com>

Examples

left("hello", 3)

[Package tmcn version 0.2-12 Index]