strcap {tmcn}	R Documentation

Mixed case capitalizing.

Description

To capitalize every first letter of a word.

Usage

strcap(string, strict = FALSE)

Arguments

string
A character vector.

strict
Whether strict.

Value

A character vector with the first letter of each word capitalized.

Author(s)

Jian Li <rweibo@sina.com>

Examples

strcap("the quick red fox jumps over the lazy brown dog")

[Package tmcn version 0.2-12 Index]