toPinyin {tmcn}	R Documentation

Convert a chinese text to pinyin format.

Description

Convert a chinese text to pinyin format.

Usage

toPinyin(string, capitalize = FALSE)

Arguments

string
A character vector.

capitalize
Whether to capitalize the first letter of each word.

Value

A character vector in pinyin format.

Author(s)

Jian Li <rweibo@sina.com>

Examples

toPinyin("the quick red fox jumps over the lazy brown dog")

[Package tmcn version 0.2-12 Index]