toTrad {tmcn}	R Documentation

Convert a Chinese text from simplified to traditional characters and vice versa.

Description

Convert a chinese text from simplified to traditional characters and vice versa.

Usage

toTrad(string, rev = FALSE)

Arguments

string
A Chinese string vector.

rev
Reverse. TRUE means traditional to simplified. Default is FALSE.

Value

Converted vectors.

Author(s)

Jian Li <rweibo@sina.com>

Examples

toTrad("hello")

[Package tmcn version 0.2-12 Index]